03_Reporter_Cluster/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2 
python-dotenv==1.0.1
pydantic==2.5.3
# Shared DB Access (if needed direct, else via Brain API)
# For V2, Reporter might read DB directly for speed
asyncpg==0.29.0
pgvector==0.2.4
Jinja2==3.1.3
playwright==1.42.0
